Prepare the raw datasets and save them to the git repository

The raw data is written to the git repository. All changes are always staged and committed. The commit is pushed when both username and password are provided.

prepare_dataset(
  scheme_id,
  raw_repo,
  result_channel,
  walloon_repo,
  flemish_channel,
  develop = FALSE,
  verbose = TRUE,
  first_year = 1992,
  latest_year = as.integer(format(Sys.time(), "%Y"))
)

Arguments

scheme_id

the id of the scheme

raw_repo

a git_repository object to write the output to

result_channel

An open RODBC connection to the results database

walloon_repo

a git_repository object to the Walloon source data

flemish_channel

a DBI connection to the Flemish database

develop

logical. Indicate whether the use the development database TRUE or the production database.

verbose

Display a progress bar when TRUE (default)

first_year

first winter to import. defaults to 1992

latest_year

latest winter to import. Winter 2019 is defined as 2018-10-01 until 2019-03-31. Defaults the winter prior to last firth of July. 2019-06-30 becomes 2018, 2019-07-01 becomes 2019.

Examples

if (FALSE) {
 prepare_dataset()
}