Read the Flemish observations from a species

The selection uses several constraints. The user defined constraints are first.winter and species.covered. Following constraints are imposed at the same time as the user defined constraints:

The observation is validated.

The observation is a mid monthly observation.

The observation is not 'unobserved'.

The location is 'active' and the observation date is within the 'active' period of the location.

The observation is between October and March.

Usage

read_observation(species_id, first_year, latest_year, flemish_channel)

Arguments

species_id

The id of the species

first_year

first winter to import. defaults to 1992

latest_year

latest winter to import. Winter 2019 is defined as 2018-10-01 until 2019-03-31. Defaults the winter prior to last firth of July. 2019-06-30 becomes 2018, 2019-07-01 becomes 2019.

flemish_channel

a DBI connection to the Flemish database

Value

A data.frame with observations. The function only returns observations based on parent locations. Following rules apply for each combination of parent location, year and month.

If the parent location has an observation, then return this observation.

If some child location have observation, then return the sum of counts.

If neither of the parent or child location has observations, then the counts are missing.

Complete = 1 indicates that the entire location was surveyed. In case of child locations, all child locations were surveyed in full. Complete = 0 indicates the location was only partially surveyed.